cgotraceback: add docs about which libunwind versions work
This package's tests pass with the built-in libunwind on macos x86-64
and arm64, and using the nongnu libunwind on x86-64. The nongnu version
doesn't seem to work from my tests on arm64 (at least, the latest commit
as of this commit doesn't work on Ubuntu Focal on an Amazon Graviton
host). However, the LLVM libunwind version does work.