cgotraceback: block SIGPROF in cgo_context and cgo_traceback
This is a kind of a big-hammer solution to #1. Blocks SIGPROF for the
duration of cgo_context and cgo_traceback to prevent the CPU profiler
from interrupting either of them (which would involve calling
cgo_traceback). The downside of this solution is that those two
callbacks won't show up on a CPU profile.

Could alternatively consider something like this:

	https://github.com/alk/unwind_safeness_helper

But that might not solve the problem for both versions of libunwind.

Fixes #1