cgotraceback: make cgo_symbolize thread-safe
I can't find any documentation guaranteeing that the libdwfl functions
used in cgo_symbolize are safe, and I've observed crashes when multiple
goroutines are concurrently creating symbolic stack traces (for
profiling) that are fixed with a lock. So I'm proactively adding a mutex
to protect cgo_symbolizer from concurrent calls.