cgotraceback: try to use unw_backtrace if it's available
The "nongnu" libunwind has the unw_backtrace convenience function that's
actually substantially faster than setting up the cursor/context, let
alone unwinding manually. So we use that if we can. I did it in two
separate files (one for macos, one for linux) because __attribute__
((weak)) doesn't seem to work the same way on macos as it does on linux.
We could use the '-U' linker flag on macos (to allow unw_backtrace to be
udefined), but that flag isn't allowed by default for cgo.