cgotraceback: add a context overhead benchmark
Whatever we do in the "context" callback will add cost to every C->Go
call, regardless of whether we actually collect a call stack. So it
needs to be as efficient of possible. Add a benchmark to compare
relative difference in latency with and without the context enabled
(via a new internal toggle to enable/disable the callbacks)