cgotraceback: use async-profiler DWARF unwinding implementation
The async-profiler project[0] contains a fairly simple and fast DWARF
call stack unwinding implementation, which additionally purports to be
signal-safe. The implementation reads unwinding information from
libraries ahead of time and builds lookup tables. When actually
unwinding a call stack, this information is only read and not modified.
Only a subset of the full DWARF call frame information is used. This
signal-safety and speed comes at the expense of additional memory usage.

[0] See https://github.com/jvm-profiling-tools/async-profiler