cgotraceback: use ucontext for unwinding from signal handler
When the traceback callback is called from a signal handler (for CPU
profiling), the ucontext_t value passed in through the sig_context field
of the can be used to recover the state of registers from the stack
that's being profiled. walkDwarf can use this to start unwinding from
the stack that's being profiled, so this is what we need to use if it's
available.

Also changed the "bottom" value in walkDwarf, which I guess is supposed
to guard against running of the end of the stack, to point to the stack
we're profiling. I noticed that when profiling a C function, the
"bottom" value would be an address on a Go stack even though the thing
we're profiling is on an OS stack. This caused unwinding to stop after
one frame due to failing the "if (sp >= bottom)" check.