cgotraceback: only partially unwind in async_cgo_context
We can fully unwind in async_cgo_traceback, when a stack is actually
requested. This reduces the overhead of async_cgo_context by quite a
bit. This was accomplished by refactoring the stack walking code to make
use of the StackContext struct, which can be populated with the PC, SP,
and FP beforehand and used later. Then async_cgo_context computes this
information, advances a few frames, and saves it. The information is
retrieved later when a stack has been requested.

The old benchmark results:

	BenchmarkContext/baseline-2         	 5677131	       208.5 ns/op
	BenchmarkContext/enabled-2          	 2233023	       532.0 ns/op

Compare to the new results:

	BenchmarkContext/baseline-2         	 5627118	       212.4 ns/op
	BenchmarkContext/enabled-2          	 4689048	       258.9 ns/op