cgotraceback: skip over possible faults during unwinding
It is possible for unwinding to lead to bad memory accesses, for example
if there's no DWARF information and we fall back to frame pointer
unwinding for a function which wasn't compiled with frame pointers. To
guard against such bad accesses, start using the safeAccess code,
together with a custom SIGSEGV handler, to skip over bad accesses if
they happen within the unwinding code.

I restored the original code from async-profiler, and tweaked it a bit
to hopefully make it a little more clearwhat it's doing.