cgotraceback: track last instruction unwinding info is good for
The FrameDesc struct tracks the first instruction the unwinding
information is good for, but not the last one. It's implicitly assumed
that the unwinding info is the same for every instruction up until the
first instruction for the following FrameDesc. But there can be holes in
the actual unwinding information, such as for Go programs (which don't
populate .eh_frame info).

This commit adds a (kind of hacky) additional field + check to see that
the chosen FrameDesc for unwinding actually covers the instruction.

Updates #4